Added comments and renamed componenets
