+++ UserSettings/EditorUserSettings.asset
@@ -21,6 +21,9 @@ EditorUserSettings:
       value: 224247031146466e0307183b370f6b365932002b21382a353e663f3aece12621c0e839e4c53527083c11fb0f073b0f3abc1f0702e212
       flags: 0
     RecentlyUsedScenePath-5:
+      value: 22424703114646680e0b0227036c7b151b180b651c242729283b3e3cecf42637eeeb33fbd33f2c287717e1351027
+      flags: 0
+    RecentlyUsedScenePath-6:
       value: 22424703114646680e0b0227036c7b151b180b650d3a2b31393c0f36c6e53937acf238e0f323
       flags: 0
     vcSharedLogLevel:
